+++ tests/unit/test_orgs.py
@@ -3,7 +3,7 @@ import pytest
 from github3 import GitHubError
 from github3.orgs import Organization
 
-from .helper import UnitIteratorHelper
+from .helper import UnitHelper, UnitIteratorHelper
 
 
 def url_for(path=''):
@@ -11,6 +11,31 @@ def url_for(path=''):
     if path:
         path = '/' + path.strip('/')
     return 'https://api.github.com/orgs/hapy' + path
+
+
+class TestOrganization(UnitHelper):
+    described_class = Organization
+    example_data = {
+        'login': 'hapy',
+        'id': 1,
+        'url': 'https://api.github.com/orgs/hapy',
+        'avatar_url': 'https://github.com/images/error/octocat_happy.gif',
+        'name': 'github',
+        'company': 'GitHub',
+        'blog': 'https://github.com/blog',
+        'location': 'San Francisco',
+        'email': 'octocat@github.com',
+        'public_repos': 2,
+        'public_gists': 1,
+        'followers': 20,
+        'following': 0,
+        'html_url': 'https://github.com/hapy',
+        'created_at': '2008-01-14T04:33:35Z',
+        'type': 'Organization'
+    }
+
+    def test_add_member(self):
+        """Show that an authenticated user can add a member to an org."""
 
 
 class TestOrganizationIterator(UnitIteratorHelper):